Split simorgh bumping scripts
